chore: move grpc tests into examples folder
skip running grpc on 3.6 due to protobuf error
3.6 passing locally on macos and on cirrus cli linux
