+++ examples/grpc/area_calculator_server.py
