+++ examples/grpc/requirements.txt
@@ -1,4 +1,4 @@
 grpcio==1.48.2; python_version < '3.7'
 grpcio==1.56.2; python_version >= '3.7'
 grpcio-tools==1.48.2; python_version < '3.7'
-grpcio-tools==1.56.2; python_version >= '3.7
+grpcio-tools==1.56.2; python_version >= '3.7'
