+++ examples/grpc/run_pytest.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+set -o pipefail
+
+pytest
