+++ requirements_dev.txt
@@ -26,7 +26,7 @@ httpx==0.22.0; python_version < '3.7'
 httpx==0.23.3; python_version >= '3.7'
 cffi>=1.14.6
 pytest-httpserver==1.0.1
-grpcio==1.48.2; python_version < '3.7'
-grpcio==1.56.2; python_version >= '3.7'
-grpcio-tools==1.48.2; python_version < '3.7'
-grpcio-tools==1.56.2; python_version >= '3.7'
+# grpcio==1.48.2; python_version < '3.7'
+# grpcio==1.56.2; python_version >= '3.7'
+# grpcio-tools==1.48.2; python_version < '3.7'
+# grpcio-tools==1.56.2; python_version >= '3.7'
