+++ tests/ffi/test_v3_verifier.py
@@ -1,10 +1,6 @@
-from sys import version_info
-from time import sleep
 import pytest
 from pact import VerifierV3
 import platform
-from os.path import join, dirname
-import subprocess
 from pact.ffi.verifier import VerifyStatus
 from pact.pact_exception import PactException
 
@@ -175,27 +171,6 @@ def test_local_http_v3_pact(httpserver):
         sources=["./examples/pacts/v3-http.json"],
     )
     assert VerifyStatus(result.return_code) == VerifyStatus.SUCCESS
-
-@pytest.mark.skipif(
-    version_info < (3, 7),
-    reason="https://stackoverflow.com/questions/71759248/importerror-cannot-import-name-builder-from-google-protobuf-internal")
-def test_grpc_local_pact():
-
-    grpc_server_process = subprocess.Popen(['python', 'area_calculator_server.py'],
-                                           cwd=join(dirname(__file__), '..', '..', 'examples', 'area_calculator'),
-                                           stdout=subprocess.PIPE,
-                                           stderr=subprocess.STDOUT, universal_newlines=True)
-    sleep(3)
-
-    verifier = VerifierV3(provider="area-calculator-provider-example",
-                          provider_base_url="tcp://127.0.0.1:37757",
-                          )
-    result = verifier.verify_pacts(
-        sources=["./examples/pacts/v4-grpc.json"],
-    )
-    grpc_server_process.terminate()
-    assert VerifyStatus(result.return_code) == VerifyStatus.SUCCESS
-    # TODO - Plugin success or failure not returned in logs
 
 def test_broker_consumer_version_selectors_http_v2_pact(httpserver):
     body = {"name": "Mary"}
